archival: pull out adjacent segment run
There's some business logic in archival/types to generate path names. In
the effort to change how we assign paths to objects, this moves that
musiness logical out of archival/types, so that the business logic
doesn't need to muddy such a fundamental file.
